add rex options
REX is going to be the original WHY2 successor, because I found the original version not as secure as I intended it to be. I am planning on implementing algorithm similar to AES, but with a bit of spice.

I already want to ditch this 'REX' project and continue procrastinating...